requirements/requirements.txt
fastapi~=0.109.0
uvicorn==0.27.0
numpy==1.26.3
pydantic==2.5.3
Pillow==10.1.0
tensorflow==2.15.0
tensorflow_hub